Feature/flavio

+++ src/main/java/com/api/mfsemijoias_cadastracliente/adapters/out/repository/ClienteRepositoryImpli.java
@@ -1,65 +1,60 @@
 package com.api.mfsemijoias_cadastracliente.adapters.out.repository;
 
-import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBMapper;
 import com.api.mfsemijoias_cadastracliente.domain.model.Cliente;
-import com.api.mfsemijoias_cadastracliente.ports.in.ClienteService;
-import lombok.Data;
+import com.amazonaws.services.dynamodbv2.datamodeling.DynamoDBMapper;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
-@Data
 @Repository
 public class ClienteRepositoryImpli implements ClienteRepository {
 
-    private final DynamoDBMapper dynamoDBMapper;
-    private final ClienteService.ClienteMapper clienteMapper;
-    private ClienteEntity clienteEntity;
-
+    @Autowired
+    private DynamoDBMapper dynamoDBMapper;
 
     @Override
     public Cliente save(Cliente cliente) {
-        clienteEntity = clienteMapper.toEntity(cliente);
+        ClienteEntity clienteEntity = ClienteMapper.toEntity(cliente);
         dynamoDBMapper.save(clienteEntity);
-        return cliente;
+        return ClienteMapper.toDomain(clienteEntity);
     }
 
     @Override
     public Cliente findById(Long id) {
-         clienteEntity = dynamoDBMapper.load(ClienteEntity.class, id);
-        return clienteMapper.toDomain(clienteEntity);
+        ClienteEntity clienteEntity = dynamoDBMapper.load(ClienteEntity.class, id);
+        return ClienteMapper.toDomain(clienteEntity);
     }
-
 
     @Override
     public Cliente findByEmail(String email) {
-        clienteEntity = dynamoDBMapper.load(ClienteEntity.class, email);
-        return clienteMapper.toDomain(clienteEntity);
+        ClienteEntity clienteEntity = dynamoDBMapper.load(ClienteEntity.class, email);
+        return ClienteMapper.toDomain(clienteEntity);
     }
 
     @Override
     public Cliente findByCpf(String cpf) {
-        clienteEntity = dynamoDBMapper.load(ClienteEntity.class, cpf);
-        return clienteMapper.toDomain(clienteEntity);
+        ClienteEntity clienteEntity = dynamoDBMapper.load(ClienteEntity.class, cpf);
+        return ClienteMapper.toDomain(clienteEntity);
     }
 
     @Override
     public Cliente findByNome(String nome) {
-        clienteEntity = dynamoDBMapper.load(ClienteEntity.class, nome);
-        return clienteMapper.toDomain(clienteEntity);
-        }
+        ClienteEntity clienteEntity = dynamoDBMapper.load(ClienteEntity.class, nome);
+        return ClienteMapper.toDomain(clienteEntity);
+    }
 
     @Override
     public Cliente deleteById(Long id) {
-        clienteEntity = dynamoDBMapper.load(ClienteEntity.class, id);
-        dynamoDBMapper.delete(clienteEntity);
-        return clienteMapper.toDomain(clienteEntity);
-
+        ClienteEntity clienteEntity = dynamoDBMapper.load(ClienteEntity.class, id);
+        if (clienteEntity != null) {
+            dynamoDBMapper.delete(clienteEntity);
+        }
+        return ClienteMapper.toDomain(clienteEntity);
     }
 
     @Override
     public Cliente update(Cliente cliente) {
-        clienteEntity = clienteMapper.toEntity(cliente);
-        dynamoDBMapper.save(cliente);
-        return cliente;
+        ClienteEntity clienteEntity = ClienteMapper.toEntity(cliente);
+        dynamoDBMapper.save(clienteEntity);
+        return ClienteMapper.toDomain(clienteEntity);
     }
-
 }
